requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==1.11.11
django-ckeditor==5.6.1
django-cors-headers==2.5.2
django-js-asset==1.2.2
django-redis==4.10.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
fdfs-client-py==1.2.6
idna==2.8
itsdangerous==1.1.0
kombu==4.5.0
lxml==4.3.3
mutagen==1.42.0
Pillow==6.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
qiniu==7.2.2
QQLoginTool==0.3.0
redis==3.2.1
requests==2.21.0
soupsieve==1.9
urllib3==1.24.1
vine==1.3.0
